+++ src/VM/call_stack.cpp
@@ -44,7 +44,7 @@ namespace elsa {
 
 		void CallStack::dump_stack_trace()
 		{
-			for (int i = stack_.size() - 1; i >= 0; --i)
+			for (int i = static_cast<int>(stack_.size()) - 1; i >= 0; --i)
 			{
 				const auto fi = stack_[i]->get_function_info();
 				std::wcout << fi->get_name() << L" at: " << fi->get_addr() << std::endl;
